backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
aiohttp>=3.8.0
pydantic>=2.0.0
python-multipart>=0.0.6
groq>=0.12.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
mypy>=1.10.0
